Attempt to re-organize Client class, Blinding and Verifying logic
