+++ veriphix/client.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import itertools
 import random
 from dataclasses import dataclass
 from math import ceil
@@ -21,110 +20,22 @@ from graphix.fundamentals import Plane
 from graphix.measurements import Measurement
 from graphix.ops import Ops
 from graphix.pattern import Pattern
-from graphix.pauli import Pauli
-from graphix.sim.statevec import State, Statevec, StatevectorBackend
+from graphix.sim.statevec import Statevec
 from graphix.simulator import MeasureMethod, PrepareMethod
 from graphix.states import BasicStates
 from stim import Circuit
 
-from veriphix.run import TestRun
+from veriphix.blinding import SecretDatas, Secrets
+from veriphix.verifying import (
+    ComputationRun,
+    ResultAnalysis,
+    TrappifiedScheme,
+    TrappifiedSchemeParameters,
+    create_test_runs,
+)
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
-
     from graphix.sim.base_backend import Backend
-
-
-from collections.abc import Set as AbstractSet
-
-Trap = AbstractSet[int]
-Traps = AbstractSet[Trap]
-
-
-@dataclass
-class TrappifiedScheme:
-    params: TrappifiedSchemeParameters
-    test_runs: list
-
-
-@dataclass
-class TrappifiedSchemeParameters:
-    comp_rounds: int  # nr of comp rounds
-    test_rounds: int  # nr of test rounds
-    threshold: int  # threshold (nr of allowed test rounds failure)
-
-
-@dataclass
-class ResultAnalysis:
-    nr_failed_test_rounds: int
-    computation_outcomes_count: dict[str, int]
-    quantum_output_states: dict[int, State]
-
-
-@dataclass
-class TrappifiedRun:
-    input_state: list
-    tested_qubits: list[int]
-    stabilizer: Pauli
-
-
-@dataclass
-class Secret_a:
-    a: dict[int, int]
-    a_N: dict[int, int]
-
-
-@dataclass
-class Secrets:
-    r: bool = True
-    a: bool = True
-    theta: bool = True
-
-
-@dataclass
-class SecretDatas:
-    r: dict[int, int]
-    a: Secret_a
-    theta: dict[int, int]
-
-    # NOTE: not a class method?
-    @staticmethod
-    def from_secrets(secrets: Secrets, graph: nx.Graph, input_nodes, output_nodes):
-        r = {}
-        if secrets.r:
-            # Need to generate the random bit for each measured qubit, 0 for the rest (output qubits)
-            for node in graph.nodes:
-                r[node] = np.random.randint(0, 2)  # if node not in output_nodes else 0
-
-        theta = {}
-        if secrets.theta:
-            # Create theta secret for all non-output nodes (measured qubits)
-            for node in graph.nodes:
-                theta[node] = np.random.randint(0, 8) if node not in output_nodes else 0  # Expressed in pi/4 units
-                ## TODO:
-        a = {}
-        a_N = {}
-        if secrets.a:
-            # Create `a` secret for all
-            # order is Z(theta) X |+>
-            for node in graph.nodes:
-                a[node] = np.random.randint(0, 2) if node in input_nodes else 0
-
-            # After all the `a` secrets have been generated, the `a_N` value can be
-            # computed from the graph topology
-            for i in graph.nodes:
-                a_N_value = sum([a[j] for j in graph.neighbors(i)]) % 2
-                # for j in graph.neighbors(i):
-                #     a_N_value ^= a[j]
-                a_N[i] = a_N_value
-
-        return SecretDatas(r, Secret_a(a, a_N), theta)
-
-    def blind_angle(secret_datas: SecretDatas, node: int, output_node: bool, test: bool) -> float:
-        r_value = 0 if (output_node and not test) else secret_datas.r.get(node, 0)
-        theta_value = secret_datas.theta.get(node, 0)
-        a_N_value = secret_datas.a.a_N.get(node, 0)
-        return theta_value * np.pi / 4 + np.pi * (r_value ^ a_N_value)
 
 
 @dataclass
@@ -216,10 +127,8 @@ class Client:
         self.preparation_bank = {}
         self.prepare_method = ClientPrepareMethod(self.preparation_bank)
 
-        from veriphix.run import ComputationRun
-
         self.computationRun = ComputationRun(self)
-        self.test_runs = self.create_test_runs()
+        self.test_runs = create_test_runs(client=self)
         self.trappifiedScheme = TrappifiedScheme(
             params=parameters or TrappifiedSchemeParameters(20, 20, 5), test_runs=self.test_runs
         )
@@ -276,25 +185,10 @@ class Client:
         The Client creates the qubits and blind them in its preparation_bank
         """
         for node in states_dict:
-            blinded_qubit_state = self.blind_qubit(node=node, state=states_dict[node])
+            blinded_qubit_state = SecretDatas.blind_qubit(
+                secret_datas=self.secret_datas, node=node, state=states_dict[node]
+            )
             self.preparation_bank[node] = Statevec(blinded_qubit_state)
-
-    def blind_qubit(self, node: int, state) -> None:
-        def z_rotation(theta) -> np.array:
-            return np.array([[1, 0], [0, np.exp(1j * theta * np.pi / 4)]])
-
-        def x_blind(a) -> Pauli:
-            return Pauli.X if (a == 1) else Pauli.I
-
-        theta = self.secret_datas.theta.get(node, 0)
-        a = self.secret_datas.a.a.get(node, 0)
-        single_qubit_backend = StatevectorBackend()
-        single_qubit_backend.add_nodes([0], [state])
-        if a:
-            single_qubit_backend.apply_single(node=0, op=x_blind(a).matrix)
-        if theta:
-            single_qubit_backend.apply_single(node=0, op=z_rotation(theta))
-        return single_qubit_backend.state
 
     def prepare_states(self, backend: Backend, states_dict: dict[(int, BasicStates)]) -> None:
         # Initializes the bank (all the nodes)
@@ -303,78 +197,6 @@ class Client:
         ## Except for the input! The Client creates them itself
         for node in self.input_nodes:
             self.prepare_method.prepare_node(backend, node)
-
-    def create_test_runs(self, manual_colouring: Sequence[set[int]] | None = None) -> list[TestRun]:
-        """Creates test runs according to a graph colouring according to [FK12].
-        A test run, or a Trappified Canvas, is associated to each color in the colouring.
-        For a given test run, the trap nodes are defined as being the nodes belonging to the color the run corresponds to.
-        This procedure only allows for the definition of single-qubit traps.
-        If `manual_colouring` is not specified, then a colouring is found by using `networkx`'s greedy algorithm.
-
-        Parameters
-        ----------
-        manual_colouring : Sequence[set[int]] | None, optional
-            manual colouring to use if `networkx` is bypassed, by default None.
-
-        Returns
-        -------
-        list[TestRun]
-            list of TestRun objects defining all the possible test runs to perform.
-
-        Raises
-        ------
-        ValueError
-            if the provided colouring is not an actual colouring (does not cover all nodes).
-        ValueError
-            if the provided colouring is not a proper colouring (a node belongs to more than one color).
-
-        Warnings
-        ------
-        when providing a `manual_colouring` make sure that the numbering of the nodes in the original pattern/graph and the one in the colouring are consistent.
-
-        Notes
-        ------
-        [FK12]: Fitzsimons, J. F., & Kashefi, E. (2017). Unconditionally verifiable blind quantum computation. Physical Review A, 96(1), 012303.
-        https://arxiv.org/abs/1203.5217
-        """
-
-        # Create the graph coloring
-        # Networkx output format: dict[int, int] eg {0: 0, 1: 1, 2: 0, 3: 1}
-        if manual_colouring is None:
-            coloring = nx.coloring.greedy_color(self.graph, strategy="largest_first")
-            colors = set(coloring.values())
-            nodes_by_color: dict[int, list[int]] = {c: [] for c in colors}
-            for node in sorted(self.graph.nodes):
-                color = coloring[node]
-                nodes_by_color[color].append(node)
-        else:
-            # checks that manual_colouring is a proper colouring
-            ## first check uniion is the whole graph
-            color_union = set().union(*manual_colouring)
-            if not color_union == set(self.graph.nodes):
-                raise ValueError("The provided colouring does not include all the nodes of the graph.")
-            # check that colors are two by two disjoint
-            # if sets are disjoint, empty set from intersection is interpreted as False.
-            # so one non-empty set -> one True value -> use any()
-            if any([i & j for i, j in itertools.combinations(manual_colouring, 2)]):
-                raise ValueError(
-                    "The provided colouring is not a proper colouring i.e the same node belongs to at least two colours."
-                )
-
-            nodes_by_color = {i: list(c) for i, c in enumerate(manual_colouring)}
-            colors = nodes_by_color.keys()
-
-        # Create the test runs : one per color
-        test_runs: list[TestRun] = []
-        for color in colors:
-            # 1 color = 1 test run = 1 collection of single-qubit traps
-            traps_list = [frozenset([colored_node]) for colored_node in nodes_by_color[color]]
-            traps: Traps = frozenset(traps_list)
-            test_run = TestRun(client=self, traps=traps)
-            test_runs.append(test_run)
-
-        # print(test_runs)
-        return test_runs
 
     def sample_canvas(self):
         N = self.trappifiedScheme.params.comp_rounds + self.trappifiedScheme.params.test_rounds
@@ -399,12 +221,16 @@ class Client:
         # True if Accept, False if Reject
         decision = result_analysis.nr_failed_test_rounds <= self.trappifiedScheme.params.threshold
 
-        # Compute majority vote
-        biased_outcome = [
-            k
-            for k, v in result_analysis.computation_outcomes_count.items()
-            if v >= ceil(self.trappifiedScheme.params.comp_rounds / 2)
-        ]
+        # Compute majority vote (only if classical output)
+        biased_outcome = (
+            [
+                k
+                for k, v in result_analysis.computation_outcomes_count.items()
+                if v >= ceil(self.trappifiedScheme.params.comp_rounds / 2)
+            ]
+            if self.classical_output
+            else None
+        )
         final_outcome = biased_outcome[0] if biased_outcome else None
 
         return decision, final_outcome
