+++ veriphix/verifying.py
@@ -1,20 +1,29 @@
+# override introduced in Python 3.12
 from __future__ import annotations
 
+import itertools
 from abc import ABC, abstractmethod
+from collections.abc import Sequence
+from collections.abc import Set as AbstractSet
+from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
+import networkx as nx
 from graphix.pattern import PatternSimulator
 from graphix.pauli import IXYZ, Pauli
+from graphix.sim.statevec import State
 from stim import PauliString
-
-# override introduced in Python 3.12
 from typing_extensions import override
+
+Trap = AbstractSet[int]
+Traps = AbstractSet[Trap]
+
 
 if TYPE_CHECKING:
     from graphix.sim.base_backend import Backend
     from graphix.states import State
 
-    from veriphix.client import Client, ResultAnalysis, Traps
+    from veriphix.client import Client, ResultAnalysis
 
 
 class Run(ABC):
@@ -186,3 +195,103 @@ class TestResult(RunResult):
 
     def analyze(self, result_analysis: ResultAnalysis, client: Client) -> None:
         result_analysis.nr_failed_test_rounds += sum(self.trap_outcomes.values()) > 0
+
+
+@dataclass
+class TrappifiedSchemeParameters:
+    comp_rounds: int  # nr of comp rounds
+    test_rounds: int  # nr of test rounds
+    threshold: int  # threshold (nr of allowed test rounds failure)
+
+
+@dataclass
+class TrappifiedScheme:
+    params: TrappifiedSchemeParameters
+    test_runs: list
+
+
+@dataclass
+class ResultAnalysis:
+    nr_failed_test_rounds: int
+    computation_outcomes_count: dict[str, int]
+    quantum_output_states: dict[int, State]
+
+
+@dataclass
+class TrappifiedRun:
+    input_state: list
+    tested_qubits: list[int]
+    stabilizer: Pauli
+
+
+def create_test_runs(client, manual_colouring: Sequence[set[int]] | None = None) -> list[TestRun]:
+    """Creates test runs according to a graph colouring according to [FK12].
+    A test run, or a Trappified Canvas, is associated to each color in the colouring.
+    For a given test run, the trap nodes are defined as being the nodes belonging to the color the run corresponds to.
+    This procedure only allows for the definition of single-qubit traps.
+    If `manual_colouring` is not specified, then a colouring is found by using `networkx`'s greedy algorithm.
+
+    Parameters
+    ----------
+    manual_colouring : Sequence[set[int]] | None, optional
+        manual colouring to use if `networkx` is bypassed, by default None.
+
+    Returns
+    -------
+    list[TestRun]
+        list of TestRun objects defining all the possible test runs to perform.
+
+    Raises
+    ------
+    ValueError
+        if the provided colouring is not an actual colouring (does not cover all nodes).
+    ValueError
+        if the provided colouring is not a proper colouring (a node belongs to more than one color).
+
+    Warnings
+    ------
+    when providing a `manual_colouring` make sure that the numbering of the nodes in the original pattern/graph and the one in the colouring are consistent.
+
+    Notes
+    ------
+    [FK12]: Fitzsimons, J. F., & Kashefi, E. (2017). Unconditionally verifiable blind quantum computation. Physical Review A, 96(1), 012303.
+    https://arxiv.org/abs/1203.5217
+    """
+
+    # Create the graph coloring
+    # Networkx output format: dict[int, int] eg {0: 0, 1: 1, 2: 0, 3: 1}
+    if manual_colouring is None:
+        coloring = nx.coloring.greedy_color(client.graph, strategy="largest_first")
+        colors = set(coloring.values())
+        nodes_by_color: dict[int, list[int]] = {c: [] for c in colors}
+        for node in sorted(client.graph.nodes):
+            color = coloring[node]
+            nodes_by_color[color].append(node)
+    else:
+        # checks that manual_colouring is a proper colouring
+        ## first check uniion is the whole graph
+        color_union = set().union(*manual_colouring)
+        if not color_union == set(client.graph.nodes):
+            raise ValueError("The provided colouring does not include all the nodes of the graph.")
+        # check that colors are two by two disjoint
+        # if sets are disjoint, empty set from intersection is interpreted as False.
+        # so one non-empty set -> one True value -> use any()
+        if any([i & j for i, j in itertools.combinations(manual_colouring, 2)]):
+            raise ValueError(
+                "The provided colouring is not a proper colouring i.e the same node belongs to at least two colours."
+            )
+
+        nodes_by_color = {i: list(c) for i, c in enumerate(manual_colouring)}
+        colors = nodes_by_color.keys()
+
+    # Create the test runs : one per color
+    test_runs: list[TestRun] = []
+    for color in colors:
+        # 1 color = 1 test run = 1 collection of single-qubit traps
+        traps_list = [frozenset([colored_node]) for colored_node in nodes_by_color[color]]
+        traps = frozenset(traps_list)
+        test_run = TestRun(client=client, traps=traps)
+        test_runs.append(test_run)
+
+    # print(test_runs)
+    return test_runs
